+++ dev-elements/playground/pixi/excercises/basics/text.js
@@ -2,7 +2,7 @@
 import * as PIXI from 'pixi.js';
 
 type Props = {
-  app: Object,
+  app: PIXI.Application,
 };
 export default function text({ app }: Props) {
   app.stage.addChild(generateBasicText());