+++ flow-typed/pixi.js
@@ -1,4 +1,0 @@
-export type PixiPoint = {
-  x: number,
-  y: number,
-};